// src/library/structure/HomeUnitarySection/HomeUnitarySection.types.ts
export interface HomeUnitarySectionProps {
  /**
    * Title
    */
  title?: string;
  /**
    * firstLine
    */
   firstLine?: string;
   /**
    * postcodeText
    */
   postcodeText?: string;
   /**
    * northText
    */
   northText?: string;
   /**
    * westText
    */
   westText?: string;
}